refactor: 폴더 구조 변경 및 파일 경로 수정
