+++ src/app/(auth)/login/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-const page = () => {
+const LoginPage = () => {
   return <div>login page</div>;
 };
 
-export default page;
+export default LoginPage;
