+++ src/app/(routes)/blog/[id]/page.tsx
