+++ tsconfig.json
@@ -21,9 +21,7 @@
     "baseUrl": "src",
     "paths": {
       "@/*": ["*"],
-      "@app/*": ["app/*"],
-      "@features/*": ["features/*"],
-      "@shared/*": ["shared/*"]
+      "@app/*": ["app/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
